project/requirements.txt
# KinFace-II Preprocessing Requirements
facenet-pytorch>=2.5.2
pillow>=9.0.0
numpy>=1.21.0
opencv-python>=4.5.5
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.0

# For EG3D (will be installed from eg3d/requirements.txt separately)
